Automatically import dev client when starting Keycloak server
