+++ js/apps/keycloak-server/package.json
@@ -1,13 +1,13 @@
 {
   "name": "keycloak-server",
+  "type": "module",
   "scripts": {
-    "start": "./scripts/start-server.mjs",
-    "import-client": "wireit",
+    "start": "wireit",
     "clear-data": "rm -r ./server/data"
   },
   "wireit": {
-    "import-client": {
-      "command": "./scripts/import-client.mjs",
+    "start": {
+      "command": "./scripts/start-server.js",
       "dependencies": [
         "../../libs/keycloak-admin-client:build"
       ]
